Moved config logic to separate ConfigHelper class
