--- src/main/java/com/google/alpollo/ConfigInfo.java
@@ -1,9 +1,0 @@
-package com.google.alpollo;
-
-public class ConfigInfo {
-    private String projectID;
-
-    public String getProjectID() {
-        return projectID;
-    }
-}
